pyproject.toml
[tool.poetry]
name = "prod-be"
version = "0.1.0"
description = ""
authors = ["nikedsb <ikedanao1230sbsnkr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pre-commit = "^4.0.1"
python-dotenv = "^1.0.1"
firebase-admin = "^6.5.0"
ruff = "^0.7.0"
django-cors-headers = "^4.5.0"
django-sslserver = "^0.22"
django-extensions = "^3.2.3"
werkzeug = "^3.0.4"
pyopenssl = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["D1","E", "F", "W", "B", "I"]
ignore = ["D100"]

fixable = ["ALL"]

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.ruff.isort]
force-single-line = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
